Refactor in preparation of future JAX-RS based BLS. WIP.
Mostly this involves isolating usage of Servlet-specific
stuff like HttpServlet[Request|Response]. Stuff that used to
live in the BlackLabServer (servlet) class was moved to the new
BlsMain singleton.

Also, request types should each have their own class instead
of relying on WebserviceParams (a big bag of URL parameters)
everywhere.

In the proxy subproject there's now different backends: a ProxyBackend
that still proxies a Solr instance, and a BlacklabBackend that doesn't
proxy, just implements the BLS API using the BlackLab libs. The corpus
info page is the only one working right now, but this shows how the
rest of the endpoints can be implemented as well.

This is a work in progress. Old BLS remains fully functional.
Merging now to avoid future merge conflicts. We'll keep chipping away
at this when there's time.

Squashed commit of the following:

commit 8abe5cd4f3fd389f7e1ee01c26dee58645dd1b54
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Tue Dec 2 14:57:08 2025 +0100

    Handle new /corpora/... URLs.

commit 3db93c06837a3fe4bdf17c050a74da2be4ceef89
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Mon Dec 1 14:03:43 2025 +0100

    API version.

commit 8b7c768f2f10c3c5ff4440aad65a9f0aea4f7a55
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Mon Dec 1 12:10:35 2025 +0100

    Corpus info response largely works.

commit 477bb5eb0b686ff3f3508bddf66a538c285b2f4e
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Fri Nov 28 12:06:22 2025 +0100

    First new BLS response, WIP.

commit 1a2a42d978a50a4eca35a6ccf4176fccfe0017f6
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Thu Nov 27 15:41:35 2025 +0100

    WIP add blacklab backend to proxy.

+++ proxy/jaxb/src/main/java/org/ivdnt/blacklab/proxy/helper/MapAdapterFieldValues.java
@@ -13,12 +13,12 @@ import jakarta.xml.bind.annotation.adapters.XmlAdapter;
 /**
  * Helps us to (de)serialize the map of fieldValues.
  */
-public class MapAdapterFieldValues extends XmlAdapter<MapAdapterFieldValues.WrapperFieldValues, Map<String, Integer>> {
+public class MapAdapterFieldValues extends XmlAdapter<MapAdapterFieldValues.WrapperFieldValues, Map<String, Long>> {
     @Override
-    public WrapperFieldValues marshal(Map<String, Integer> m) {
+    public WrapperFieldValues marshal(Map<String, Long> m) {
         WrapperFieldValues wrapper = new WrapperFieldValues();
         List<FieldValueFreq> elements = new ArrayList<>();
-        for (Map.Entry<String, Integer> property : m.entrySet()) {
+        for (Map.Entry<String, Long> property : m.entrySet()) {
             elements.add(new FieldValueFreq(property.getKey(), property.getValue()));
         }
         wrapper.elements = elements;
@@ -26,8 +26,8 @@ public class MapAdapterFieldValues extends XmlAdapter<MapAdapterFieldValues.Wrap
     }
 
     @Override
-    public Map<String, Integer> unmarshal(WrapperFieldValues v) {
-        Map<String, Integer> returnval = new LinkedHashMap();
+    public Map<String, Long> unmarshal(WrapperFieldValues v) {
+        Map<String, Long> returnval = new LinkedHashMap();
         for (FieldValueFreq e: v.elements) {
             returnval.put(e.text, e.freq);
         }
@@ -40,9 +40,9 @@ public class MapAdapterFieldValues extends XmlAdapter<MapAdapterFieldValues.Wrap
         public String text;
 
         @XmlValue
-        public Integer freq;
+        public Long freq;
 
-        public FieldValueFreq(String text, Integer freq) {
+        public FieldValueFreq(String text, Long freq) {
             this.text = text;
             this.freq = freq;
         }
